Add files via upload
My first project that I will be using NO tutorials walking me through the step by step process. I am attempting to create a box that will change color dependent on the A,R,G,B value, using buttons and a left click to increase the value by the set amount or decrease the value if a right click. I would love to get any and all feedback on what I have so far as this is my only project to be finished at the moment